+++ src/hooks/use-transactions-list-state.ts
@@ -1,6 +1,8 @@
 import {
+  MempoolTransactionsListResponse,
   mempoolTransactionsListState,
   OptionalTransactionAddress,
+  TransactionsListResponse,
   transactionsListState,
 } from '@store/transactions';
 import { DEFAULT_LIST_LIMIT_SMALL, IS_DEV } from '@common/constants';
@@ -8,6 +10,7 @@ import { useMemo } from 'react';
 import { useAtomDevtools } from '@common/hooks/use-atom-devtools';
 import { useInfiniteQueryAtom } from 'jotai-query-toolkit';
 import { DEFAULT_TX_FILTER_TYPES } from '@store/recoil/filter';
+import { InfiniteData } from 'react-query';
 
 export function useTransactionsListState(
   limit = DEFAULT_LIST_LIMIT_SMALL,
@@ -19,7 +22,7 @@ export function useTransactionsListState(
     return transactionsListState([limit, types, options]);
   }, [limit, types, options]);
   if (IS_DEV) useAtomDevtools(anAtom as any);
-  return useInfiniteQueryAtom(anAtom);
+  return useInfiniteQueryAtom<InfiniteData<TransactionsListResponse> | undefined>(anAtom);
 }
 
 export function useMempoolTransactionsListState(
@@ -28,5 +31,5 @@ export function useMempoolTransactionsListState(
 ) {
   const anAtom = mempoolTransactionsListState([limit, options]);
   if (IS_DEV) useAtomDevtools(anAtom as any);
-  return useInfiniteQueryAtom(anAtom);
+  return useInfiniteQueryAtom<InfiniteData<MempoolTransactionsListResponse> | undefined>(anAtom);
 }