+++ src/pages/txid/[txid].tsx
@@ -13,6 +13,7 @@ import type { TxPageQueryProps } from '@common/page-queries/txid';
 import { Meta } from '@components/meta-head';
 import { TxNotFound } from '@components/tx-not-found';
 import { InView } from '@store/currently-in-view';
+import { useChainModeEffect } from '@common/hooks/use-chain-mode';
 
 interface TransactionPageProps {
   inView: InView;
@@ -21,6 +22,7 @@ interface TransactionPageProps {
 }
 
 const TransactionPage: NextPage<TransactionPageProps> = ({ error, isPossiblyValid }) => {
+  useChainModeEffect();
   if (error)
     return (
       <>