+++ src/store/common.ts
@@ -1,4 +1,4 @@
-export function getNextPageParam(options: { limit: number; offset: number; total: number }) {
+export function getNextPageParam(options?: { limit: number; offset: number; total: number }) {
   if (!options) return 0;
   const { limit, offset, total } = options;
   const sum = offset + limit;